+++ src/son_editor/apis/platformsapi.py
@@ -5,9 +5,9 @@ Created on 18.07.2016
 """
 from flask_restplus import Resource, Namespace
 
-from son_editor.app.util import prepare_response
 from son_editor.app.constants import WORKSPACES, PLATFORMS
-from son_editor.platforms import platformsimpl
+from son_editor.app.util import prepare_response
+from son_editor.impl import platformsimpl
 
 namespace = Namespace(WORKSPACES + '/<int:ws_id>/' + PLATFORMS, description="Platform Resources")
