--- src/son_editor/catalogues/__init__.py
@@ -1,5 +1,0 @@
-from . import cataloguesapi
-
-
-def init(api):
-    api.add_namespace(cataloguesapi.namespace)
