+++ src/son_editor/impl/projectsimpl.py
@@ -12,9 +12,9 @@ from subprocess import Popen, PIPE
 from son_editor.app.database import db_session
 from son_editor.app.exceptions import NotFound, NameConflict
 from son_editor.app.util import CONFIG, rreplace
+from son_editor.impl.usermanagement import get_user
 from son_editor.models.project import Project
 from son_editor.models.workspace import Workspace
-from son_editor.users.usermanagement import get_user
 
 if CONFIG['testing']:
     WORKSPACES_DIR = tempfile.mkdtemp()
