--- src/son_editor/misc/__init__.py
@@ -1,5 +1,0 @@
-from . import misc
-
-
-def init(api):
-    api.add_namespace(misc.namespace)
