--- src/son_editor/platforms/__init__.py
@@ -1,5 +1,0 @@
-from . import platformsapi
-
-
-def init(api):
-    api.add_namespace(platformsapi.namespace)
