--- src/son_editor/projects/__init__.py
@@ -1,5 +1,0 @@
-from . import projectsapi
-
-
-def init(api):
-    api.add_namespace(projectsapi.namespace)
