--- src/son_editor/services/__init__.py
@@ -1,7 +1,0 @@
-from . import servicesapi
-
-
-def init(api):
-    api.add_namespace(servicesapi.proj_namespace)
-    api.add_namespace(servicesapi.cata_namespace)
-    api.add_namespace(servicesapi.plat_namespace)
