--- src/son_editor/workspaces/__init__.py
@@ -1,5 +1,0 @@
-from . import workspacesapi
-
-
-def init(api):
-    api.add_namespace(workspacesapi.namespace)
